Break up admin::rest_api::proposals_read module
This divides the module into two resource modules and two associated
actix modules. Also moves make_*_resource functions.
