+++ libsplinter/src/admin/rest_api/actix/proposals_read.rs
@@ -10,65 +10,34 @@
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
-// limitations under the License.
+
+use actix_web::{error::BlockingError, web, Error, HttpRequest, HttpResponse};
+use futures::{future::IntoFuture, Future};
 use std::collections::HashMap;
 
-use crate::actix_web::{error::BlockingError, web, Error, HttpRequest, HttpResponse};
 use crate::admin::messages::CircuitProposal;
-use crate::admin::rest_api::error::ProposalRouteError;
 use crate::admin::service::AdminCommands;
-use crate::futures::{future::IntoFuture, Future};
-use crate::rest_api::paging::{get_response_paging_info, Paging, DEFAULT_LIMIT, DEFAULT_OFFSET};
+use crate::protocol;
+use crate::rest_api::paging::{get_response_paging_info, DEFAULT_LIMIT, DEFAULT_OFFSET};
+use crate::rest_api::{Method, ProtocolVersionRangeGuard, Resource};
 
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
-pub struct ListProposalsResponse {
-    data: Vec<CircuitProposal>,
-    paging: Paging,
-}
+use super::super::error::ProposalRouteError;
+use super::super::resources::proposals_read::ListProposalsResponse;
 
-pub fn fetch_proposal<A: AdminCommands + Clone + 'static>(
-    request: HttpRequest,
-    admin_commands: web::Data<A>,
-) -> Box<dyn Future<Item = HttpResponse, Error = Error>> {
-    let circuit_id = request
-        .match_info()
-        .get("circuit_id")
-        .unwrap_or("")
-        .to_string();
-    Box::new(
-        web::block(move || {
-            let proposal = admin_commands
-                .fetch_proposal(circuit_id.clone())
-                .map_err(|err| ProposalRouteError::InternalError(err.to_string()))?;
-            if let Some(proposal) = proposal {
-                let proposal = CircuitProposal::from_proto(proposal)
-                    .map_err(|err| ProposalRouteError::InternalError(err.to_string()))?;
-
-                Ok(proposal)
-            } else {
-                Err(ProposalRouteError::NotFound(format!(
-                    "Unable to find proposal: {}",
-                    circuit_id
-                )))
-            }
+pub fn make_list_proposals_resource<A: AdminCommands + Clone + 'static>(
+    admin_commands: A,
+) -> Resource {
+    Resource::build("admin/proposals")
+        .add_request_guard(ProtocolVersionRangeGuard::new(
+            protocol::ADMIN_LIST_PROPOSALS_PROTOCOL_MIN,
+            protocol::ADMIN_PROTOCOL_VERSION,
+        ))
+        .add_method(Method::Get, move |r, _| {
+            list_proposals(r, web::Data::new(admin_commands.clone()))
         })
-        .then(|res| match res {
-            Ok(proposal) => Ok(HttpResponse::Ok().json(proposal)),
-            Err(err) => match err {
-                BlockingError::Error(err) => match err {
-                    ProposalRouteError::InternalError(_) => {
-                        error!("{}", err);
-                        Ok(HttpResponse::InternalServerError().into())
-                    }
-                    ProposalRouteError::NotFound(err) => Ok(HttpResponse::NotFound().json(err)),
-                },
-                _ => Ok(HttpResponse::InternalServerError().into()),
-            },
-        }),
-    )
 }
 
-pub fn list_proposals<A: AdminCommands + Clone + 'static>(
+fn list_proposals<A: AdminCommands + Clone + 'static>(
     req: HttpRequest,
     admin_commands: web::Data<A>,
 ) -> Box<dyn Future<Item = HttpResponse, Error = Error>> {
